+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Extension: Fixed typescript handling.
 - Tailwind: Revert back to automatic delivery of css, and allow for one index.css file.
 - Template: Don't try to export type only imports.
+- Server: Added scheduler
 
 ## [0.4.3] = 2024-04-05
 ### Fixed